add feature: sqlla supports tinygo. for the edge worker services
